requirements.txt
Django~=2.2.4
django-rest-swagger==2.2.0 
djangorestframework==3.11.0
PyYAML==5.3.1
django-rest-swagger-swaggerdoc==1.0.post3